google-cloud-nio: CloudStorageReadChannel: fix critical bug that could cause the channel position to overflow
This fixes a critical bug in CloudStorageReadChannel in which the channel position
(a long) was getting down-cast to int in CloudStorageReadChannel.innerOpen(). This
could cause the channel position to become negative or a wildly incorrect positive
value if the channel position was greater than MAX_INT.

In practice, this bug would only manifest itself if the channel was constructed with
a non-zero initial position, or if the channel was reopened in response to an error.

Included a test case that fails without the fix, and passes with the fix.

+++ google-cloud-contrib/google-cloud-nio/src/main/java/com/google/cloud/storage/contrib/nio/CloudStorageReadChannel.java
@@ -82,7 +82,7 @@ final class CloudStorageReadChannel implements SeekableByteChannel {
   private void innerOpen() throws IOException {
     this.channel = gcsStorage.reader(file);
     if (position > 0) {
-      channel.seek((int) position);
+      channel.seek(position);
     }
   }
